create new controller and model for article table
