+++ app.js
@@ -1,10 +1,11 @@
 const express = require("express");
-const { getTopics, getArticlesById } = require("./controller/topic.controller");
+const { getTopics, getArticlesById, patchArticleById } = require("./controller/topic.controller");
 const app = express();
 
 app.get("/api/topics", getTopics);
 
 app.get("/api/articles/:article_id", getArticlesById);
+app.patch("/api/articles/:article_id", patchArticleById)
 
 app.all("/*", (req, res) => {
   res.status(404).send({ msg: "Not Found" });
